requirements.txt
boto3==1.34.144
python-dotenv==1.0.1
Faker==0.7.4
pandas==2.2.2
apache-airflow==2.9.2
apache-airflow-providers-postgres==5.11.2
apache-airflow-providers-amazon==8.26.0
psycopg2-binary==2.9.9
SQLAlchemy==1.4.52
